+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 # 3.4.0
 
-- Use `prune-allowlist` instead of `prune-whitelist`. This change requires kubernetes 1.26 and higher. [#940](https://github.com/Shopify/krane/pull/940)
+- Use `prune-allowlist` instead of `prune-whitelist` for 1.26+ clusters. Clusters running 1.25 or less will continue to use `--prune-whitelist`. [#940](https://github.com/Shopify/krane/pull/940)
 
 ## 3.3.0